requirements.txt
asgiref==3.4.1
boto3==1.18.58
botocore==1.21.58
Django==3.2.8
django-storages==1.12
django-tinymce==3.3.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
